# Conflicts:
#	src/main/java/edu/depaul/cdm/se352452group4/groupProject/controller/AccountController.java
#	src/main/java/edu/depaul/cdm/se352452group4/groupProject/model/entity/Account.java
#	src/main/resources/application.properties

+++ src/main/java/edu/depaul/cdm/se352452group4/groupProject/model/entity/Transactions.java
@@ -13,7 +13,7 @@ public class Transactions implements Serializable {
 
     @Id
     @GeneratedValue
-    private Integer transactionId; // PK.
+    private Long transactionId; // PK.
     private Double transactionsTotal; // small money
     private Integer totalItems;
